+++ src/App.tsx
@@ -1,16 +1,23 @@
 import React, { useState } from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
-import { Col, Row } from "react-bootstrap";
+import { Container, Col, Row } from "react-bootstrap";
+import { useLocation } from "react-router-dom";
+import { library } from "@fortawesome/fontawesome-svg-core";
+import { fab } from "@fortawesome/free-brands-svg-icons";
+import { fas } from "@fortawesome/free-solid-svg-icons";
 import Header from "./components/Header";
 import LoanChart from "./components/LoanChart";
 import LoanForm from "./components/LoanForm";
 import LoanResult from "./components/LoanResult";
 import { computeLoanParameters } from "./utils/index";
 
+library.add(fas, fab);
+
 const App = (): JSX.Element => {
-  const [amount, setAmount] = useState(0);
-  const [term, setTerm] = useState(0);
-  const [interest, setInterest] = useState(0);
+  const params = new URLSearchParams(useLocation().search);
+  const [amount, setAmount] = useState(Number(params.get("amount") || 0));
+  const [term, setTerm] = useState(Number(params.get("term") || 0));
+  const [interest, setInterest] = useState(Number(params.get("interest") || 0));
   const { monthlyPayment, totalInterest, total } = computeLoanParameters(
     amount,
     term,
@@ -18,17 +25,20 @@ const App = (): JSX.Element => {
   );
   return (
     <div className="App">
-      <Header />
-      <div className="container">
+      <Header amount={amount} term={term} interest={interest} />
+      <Container>
         <Row>
-          <Col sm={12} md={4}>
+          <Col sm={12} md={6} lg={4} className="mt-3">
             <LoanForm
+              amount={amount}
               onAmount={setAmount}
+              term={term}
               onTerm={setTerm}
+              interest={interest}
               onInterest={setInterest}
             />
           </Col>
-          <Col>
+          <Col sm={12} md={6} lg={4} className="mt-3">
             <LoanResult
               monthlyPayment={monthlyPayment}
               amount={amount}
@@ -36,7 +46,7 @@ const App = (): JSX.Element => {
               total={total}
             />
           </Col>
-          <Col>
+          <Col lg={4} className="mt-3">
             <LoanChart
               amount={amount}
               totalInterest={totalInterest}
@@ -44,7 +54,7 @@ const App = (): JSX.Element => {
             />
           </Col>
         </Row>
-      </div>
+      </Container>
     </div>
   );
 };