+++ src/components/types.tsx
@@ -1,0 +1,7 @@
+type LoanParamsProps = {
+  amount: number;
+  term: number;
+  interest: number;
+};
+
+export default LoanParamsProps;