+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [2021.06.13]
+  - Publish to GitHub pages
+  - Use FontAwesome
+  - Improve responsiveness
+  - Leverage React Router for loan parameters share link
 
 ## [2021.06.06]
   - Initial release